chore: check if auth user is not user to delete

+++ src/controllers/user.controller.ts
@@ -1,9 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Request, Response, NextFunction } from 'express'
+import createHttpError from 'http-errors'
 
-import { userTypes } from '../types'
+import { userTypes, tokenTypes } from '../types'
 
 import { userService } from '../services'
+
+import { getTokenFromHeader } from '../utils/getTokenFromHeader'
+import { getJwtPayload } from '../utils/getJwtPayload'
 
 async function getUsers(req: Request, res: Response, next: NextFunction) {
   const users: userTypes.UserDocument[] = await userService.getUsers()
@@ -31,6 +35,12 @@ async function updateUser(req: Request, res: Response, next: NextFunction) {
 }
 
 async function deleteUser(req: Request, res: Response, next: NextFunction) {
+  const { sub }: tokenTypes.Payload = getJwtPayload(getTokenFromHeader(req))
+
+  if (sub === req.params.id) {
+    throw createHttpError(400, 'You cannot delete yourself')
+  }
+
   await userService.deleteUser(req.params.id)
   res.status(200).send()
 }
